+++ domain/employee_shift_entity.go
@@ -7,9 +7,9 @@ type EmployeeShift struct {
 	BranchUUID string                   `bson:"branch_uuid"`
 	UserUUID   string                   `bson:"user_uuid"`
 	StartTime  int64                    `bson:"start_time"`
-	StartCash  float32                  `bson:"start_cash"`
+	StartCash  float64                  `bson:"start_cash"`
 	EndTime    *int64                   `bson:"end_time"`
-	EndCash    *float32                 `bson:"end_cash"`
+	EndCash    *float64                 `bson:"end_cash"`
 	Supporters []EmployeeShiftSupporter `bson:"supporters"`
 	CreatedAt  int64                    `bson:"created_at"`
 	UpdatedAt  *int64                   `bson:"updated_at"`
@@ -32,13 +32,13 @@ type EmployeeShiftClockInData struct {
 	UserUUID       string   `bson:"user_uuid"`
 	SupportingUUID *string  `bson:"supporting_uuid"`
 	StartTime      int64    `bson:"start_time"`
-	StartCash      *float32 `bson:"start_cash"`
+	StartCash      *float64 `bson:"start_cash"`
 }
 
 type EmployeeShiftClockOutData struct {
 	UUID    string   `bson:"uuid"`
 	EndTime int64    `bson:"end_time"`
-	EndCash *float32 `bson:"end_cash"`
+	EndCash *float64 `bson:"end_cash"`
 }
 
 // -- data that are used to pass in to the repository
